Add files via upload
This code is Just for testing purpose, hence the exposure of  sensitive data, The are not real life Data, just for Testing.

The version 2 came with a lot of upgrade, changes to information are saved real time and are updated so long as the code is still running. Want To change ATM PIN, change and login with it.